+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * fix: Fix the bug that there will be white space on the edge of the item when there is only sideDivider in GridDividerItemDecoration
 * change: Now \*DividerItemDecoration useDividerAs\*Divider is true but an exception will be thrown when the corresponding divider is null
 * change: Now GridDividerItemDecoration and StaggeredGridDividerItemDecoration will throw an exception when there is sideHeaderDivider or sideFooter but no sideDivider
+* change: Now GridDividerItemDecoration and StaggeredGridDividerItemDecoration no longer allow sideHeaderDivider or sideFooterDivider and sideDivider to have different sizes and insets
 
 # v4.0.1